Fix: 카테고리, 즐겨찾기 다대다로 수정

+++ src/main/java/devkor/com/teamcback/domain/bookmark/dto/response/GetCategoryRes.java
@@ -16,7 +16,7 @@ public class GetCategoryRes {
     @Schema(description = "카테고리 메모", example = "자주 찾는 장소 모음")
     private String memo;
     @Schema(description = "카테고리의 즐겨찾기 수", example = "5")
-    private Long bookmarkCount;
+    private int bookmarkCount;
 
     public GetCategoryRes(Category category) {
         this.categoryId = category.getId();
@@ -25,7 +25,7 @@ public class GetCategoryRes {
         this.memo = category.getMemo();
     }
 
-    public GetCategoryRes(Category category, Long bookmarkCount) {
+    public GetCategoryRes(Category category, int bookmarkCount) {
         this.categoryId = category.getId();
         this.category = category.getCategory();
         this.color = category.getColor().getName();
